Extract hover
