+++ core/main.py
@@ -31,13 +31,10 @@ from .clients import (
 )
 from .events import Events
 from .documents import purge_did_change, get_document_position, initialize_document_sync, notify_did_open
-from .diagnostics import handle_diagnostics, remove_diagnostics, get_line_diagnostics, get_point_diagnostics
+from .diagnostics import handle_diagnostics, remove_diagnostics, get_line_diagnostics
 from .panels import create_output_panel
 
-
 SUBLIME_WORD_MASK = 515
-NO_HOVER_SCOPES = 'comment, constant, keyword, storage, string'
-
 
 symbol_kind_names = {
     SymbolKind.File: "file",
@@ -149,6 +146,15 @@ def handle_initialize_result(result, client, window, config):
     if settings.show_status_messages:
         window.status_message("{} initialized".format(config.name))
     didopen_after_initialize = list()
+
+
+def preserve_whitespace(contents: str) -> str:
+    """Preserve empty lines and whitespace for markdown conversion."""
+    contents = contents.strip(' \t\r\n')
+    contents = contents.replace('\t', '&nbsp;' * 4)
+    contents = contents.replace('  ', '&nbsp;' * 2)
+    contents = contents.replace('\n\n', '\n&nbsp;\n')
+    return contents
 
 
 class LspSymbolRenameCommand(sublime_plugin.TextCommand):
@@ -457,109 +463,6 @@ def start_server(server_binary_args, working_dir, env):
     except Exception as err:
         sublime.status_message("Failed to start LSP server {}".format(str(server_binary_args)))
         exception_log("Failed to start server", err)
-
-
-class HoverHandler(sublime_plugin.ViewEventListener):
-    def __init__(self, view):
-        self.view = view
-
-    @classmethod
-    def is_applicable(cls, settings):
-        syntax = settings.get('syntax')
-        return syntax and is_supported_syntax(syntax)
-
-    def on_hover(self, point, hover_zone):
-        if hover_zone != sublime.HOVER_TEXT or self.view.is_popup_visible():
-            return
-        point_diagnostics = get_point_diagnostics(self.view, point)
-        if point_diagnostics:
-            self.show_diagnostics_hover(point, point_diagnostics)
-        else:
-            self.request_symbol_hover(point)
-
-    def request_symbol_hover(self, point):
-        if self.view.match_selector(point, NO_HOVER_SCOPES):
-            return
-        client = client_for_view(self.view)
-        if client and client.has_capability('hoverProvider'):
-            word_at_sel = self.view.classify(point)
-            if word_at_sel & SUBLIME_WORD_MASK:
-                document_position = get_document_position(self.view, point)
-                if document_position:
-                    client.send_request(
-                        Request.hover(document_position),
-                        lambda response: self.handle_response(response, point))
-
-    def handle_response(self, response, point):
-        debug(response)
-        if self.view.is_popup_visible():
-            return
-        contents = "No description available."
-        if isinstance(response, dict):
-            # Flow returns None sometimes
-            # See: https://github.com/flowtype/flow-language-server/issues/51
-            contents = response.get('contents') or contents
-        self.show_hover(point, contents)
-
-    def show_diagnostics_hover(self, point, diagnostics):
-        formatted = list("{}: {}".format(diagnostic.source, diagnostic.message) for diagnostic in diagnostics)
-        formatted.append("[{}]({})".format('Code Actions', 'code-actions'))
-        mdpopups.show_popup(
-            self.view,
-            "\n".join(formatted),
-            css=".mdpopups .lsp_hover { margin: 4px; }",
-            md=True,
-            flags=sublime.HIDE_ON_MOUSE_MOVE_AWAY,
-            location=point,
-            wrapper_class="lsp_hover",
-            max_width=800,
-            on_navigate=lambda href: self.on_diagnostics_navigate(href, point, diagnostics))
-
-    def on_diagnostics_navigate(self, href, point, diagnostics):
-        # TODO: don't mess with the user's cursor.
-        # Instead, pass code actions requested from phantoms & hovers should call lsp_code_actions with
-        # diagnostics as args, positioning resulting UI close to the clicked link.
-        sel = self.view.sel()
-        sel.clear()
-        sel.add(sublime.Region(point, point))
-        self.view.run_command("lsp_code_actions")
-
-    def show_hover(self, point, contents):
-        formatted = []
-        if not isinstance(contents, list):
-            contents = [contents]
-
-        for item in contents:
-            value = ""
-            language = None
-            if isinstance(item, str):
-                value = item
-            else:
-                value = item.get("value")
-                language = item.get("language")
-            if language:
-                formatted.append("```{}\n{}\n```".format(language, value))
-            else:
-                formatted.append(value)
-
-        mdpopups.show_popup(
-            self.view,
-            preserve_whitespace("\n".join(formatted)),
-            css=".mdpopups .lsp_hover { margin: 4px; } .mdpopups p { margin: 0.1rem; }",
-            md=True,
-            flags=sublime.HIDE_ON_MOUSE_MOVE_AWAY,
-            location=point,
-            wrapper_class="lsp_hover",
-            max_width=800)
-
-
-def preserve_whitespace(contents: str) -> str:
-    """Preserve empty lines and whitespace for markdown conversion."""
-    contents = contents.strip(' \t\r\n')
-    contents = contents.replace('\t', '&nbsp;' * 4)
-    contents = contents.replace('  ', '&nbsp;' * 2)
-    contents = contents.replace('\n\n', '\n&nbsp;\n')
-    return contents
 
 
 class SignatureHelpListener(sublime_plugin.ViewEventListener):
